pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "basic-grpc-service-python"
version = "0.1.0"
description = "Python gRPC server/client for BasicService"
requires-python = ">=3.10"
dependencies = [
    "grpcio",
    "grpcio-reflection",
    "grpcio-health-checking",
    "protobuf~=6.31.1",
    "types-protobuf~=6.30",
    "pyright",
    "protobuf-init",
    "python-json-logger",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]
docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-section-index",
]

[tool.setuptools]
package-dir = { "" = "sdk" } # map top-level packages to ./sdk

[tool.setuptools.packages.find]
where = ["sdk"]
include = ["*"]

[tool.pyright]
exclude = ["**/node_modules", "**/__pycache__", "sdk/**"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
minversion = "6.0"
